# backend/internal/handler/openapi.yaml
openapi: 3.0.3
info:
  title: 影策 HTTP API
  version: "1.0.0"
  description: |
    路径前缀是 `/api`。前后端一起发布，不按路径做 API 版本。

    业务 JSON 信封为 `{ code, data, msg, reason }`。`code === 0` 表示成功；失败时 HTTP status 与
    `code` 都表达失败。`code` 默认等于 HTTP 状态；同一状态下的细分原因使用 `status*100+n`
    （例如 40301 配额、42901 限流）。机器判断请用 `reason`，不要解析 `msg`。

    TypeScript 调用面仍是 `web/src/services/api` 的 `http` 模块，不从本规范生成客户端。
servers:
  - url: /api
security:
  - sessionCookie: []
...
paths:
...
  /canvas-projects/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string }
    get:
      tags: [canvas]
      summary: 读取当前用户画布内容和 revision
      responses:
        "200":
          description: 信封 data.project 包含完整画布及服务端 revision
        "404":
          description: 画布不存在或不属于当前用户
    put:
      tags: [canvas]
      summary: 按 revision 保存画布内容
      description: 新建传 revision 0，更新传读取到的版本；成功原子加 1。忽略客户端 viewport 和 remoteContentHash，updatedAt 由服务端生成。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [project]
              properties:
                project:
                  type: object
                  required: [id, revision]
                  properties:
                    id: { type: string }
                    revision: { type: integer, format: int64, minimum: 0 }
                  additionalProperties: true
      responses:
        "200":
          description: 信封 data.project 返回摘要及递增后的 revision
        "400":
          description: 请求、版本或素材引用不合法
        "409":
          description: 版本或资源状态冲突，原内容未被覆盖；保留草稿后加载最新版
        "428":
          description: 缺少 revision，需刷新客户端并重新读取
  /canvas-projects/{id}/history:
    get:
      tags: [canvas]
      summary: 读取当前画布版本与最近 20 份历史摘要
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: 信封 data 包含 currentRevision 和 snapshots；每项含 id、revision、title、nodeCount、connectionCount、payloadBytes、reason、contentUpdatedAt、createdAt，不含正文
        "404":
          description: 画布不存在或无权访问
  /canvas-projects/{id}/history/{snapshotId}:
    get:
      tags: [canvas]
      summary: 按需读取历史正文
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
        - in: path
          name: snapshotId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: 信封 data 包含 snapshot 摘要和 project 正文
        "404":
          description: 画布或快照不存在、已过期或无权访问
  /canvas-projects/{id}/history/{snapshotId}/restore:
    post:
      tags: [canvas]
      summary: 备份当前内容并将历史内容恢复为新版本
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
        - in: path
          name: snapshotId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [revision]
              properties:
                revision: { type: integer, format: int64, minimum: 1 }
      responses:
        "200":
          description: 信封 data.project 返回恢复后的摘要及新 revision；当前 ID、归属与项目关联保留
        "404":
          description: 画布或快照不存在、已过期或无权访问
        "409":
          description: 当前版本或素材状态已变化，恢复未执行
        "428":
          description: 缺少当前 revision
...
  /resources:
    get:
      tags: [resources]
      summary: 当前用户资源列表
      responses:
        "200":
          $ref: "#/components/responses/Envelope"
    post:
      tags: [resources]
      summary: 上传资源文件
      responses:
        "200":
          $ref: "#/components/responses/Envelope"
        "403":
          description: 配额用尽（code=40301, reason=quota_exceeded）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /resources/{id}/file:
    get:
      tags: [resources]
      summary: 读取资源二进制（非 JSON 信封）
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: 文件流
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
...
components:
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: open_ai_canvas_session
  schemas:
...
    Envelope:
      type: object
      required: [code, msg]
      properties:
        code:
          type: integer
          description: 0 成功；失败时默认等于 HTTP 状态，或 status*100+n 的细分码
        data: {}
        msg:
          type: string
        reason:
          type: string
          description: 仅失败时出现。稳定机器可读原因。
    ErrorEnvelope:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            code:
              type: integer
              enum: [400, 401, 403, 404, 409, 429, 500, 502, 503, 504, 40301, 40901, 42901]
            data:
              nullable: true
            reason:
              type: string
              enum:
                - invalid_argument
                - unauthorized
                - forbidden
                - not_found
                - conflict
                - failed_precondition
                - quota_exceeded
                - rate_limited
                - unavailable
                - timeout
                - internal
                - bad_gateway
                - model_capability_not_supported
                - model_price_not_configured
                - model_route_unavailable
                - provider_request_failed
                - model_catalog_mismatch
                - invalid_model_selection
  responses:
...
    Envelope:
      description: 业务成功信封
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Envelope"
    Error:
      description: 业务失败信封
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"